+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+- (Experimental) Aggregation Extension endpoint /aggregate
 - Added pre-hook and post-hook Lambda examples
 - POST /collections endpoint to create collections